pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shard_core"
version = "0.37.3"
description = "Core software stack that manages all aspects of a Shard"
readme = "README.md"
license-files = ["LICENSE.md"]
authors = [
    { name = "Max von Tettenborn", email = "max.t@freeshard.net" }
]
urls = { "Homepage" = "https://github.com/FreeshardBase/shard_core" }
requires-python = ">=3.13"
dependencies = [
    "pydantic-settings>=2,<3",
    "psycopg[binary]",
    "psycopg_pool",
    "yoyo-migrations",
    "uvicorn",
    "fastapi[standard]",
    "websockets",
    "pyjwt",
    "pydantic>=2,<3",
    "Jinja2",
    "pyyaml",
    "docker",
    "python-gitlab",
    "cachetools",
    "blinker",
    "requests",
    "aiohttp",
    "cryptography",
    "requests-http-signature",
    "aiozipstream",
    "email_validator",
    "croniter",
    "bitstring",
    "azure-storage-blob",
    "asgi-lifespan==2.*",
    "python-multipart",
    "aiofiles",
    "httpx",
]

[project.optional-dependencies]
dev = [
    "setuptools",
    "ruff",
    "pytest",
    "pytest-docker",
    "pytest-mock",
    "pytest-asyncio",
    "yappi",
    "responses",
    "aioresponses",
    "datamodel-code-generator[http]"
]

[tool.setuptools]
py-modules = []
package-data = { "shard_core" = ["config.toml", "data/*"] }

[tool.ruff]
line-length = 120
lint.ignore = ["E712"]

[tool.pytest.ini_options]
log_cli = true
asyncio_mode = "auto"
# https://github.com/pytest-dev/pytest-asyncio/issues/924
asyncio_default_fixture_loop_scope = "function"
